Fixed model structure naming Separated yamls Refactored everything
