+++ developer_configuration.yml
@@ -1,0 +1,61 @@
+available_models:
+  - name: OpenAI
+    class: llmproxy.provider.openai.chatgpt.OpenAI
+    models:
+      - name: gpt-3.5-turbo-1106
+        cost_per_token_input: 0.0000010
+        cost_per_token_output: 0.0000020
+      - name: gpt-3.5-turbo-instruct
+        cost_per_token_input: 0.0000015
+        cost_per_token_output: 0.0000020
+      - name: gpt-4
+        cost_per_token_input: 0.00003
+        cost_per_token_output: 0.00006
+      - name: gpt-4-32k
+        cost_per_token_input: 0.00006
+        cost_per_token_output: 0.00012
+
+  - name: Llama2
+    class: llmproxy.provider.huggingface.llama2.Llama2
+    models:
+      - name: Llama-2-7b-chat-hf
+        cost_per_hour: 0.06
+      - name: Llama-2-13b-chat-hf
+        cost_per_hour: 0.06
+      - name: Llama-2-70b-chat-hf
+        cost_per_hour: 0.06
+
+  - name: Mistral
+    class: llmproxy.provider.huggingface.mistral.Mistral
+    models:
+      - name: Mistral-7B-v0.1
+        cost_per_hour: 0.06
+      - name: Mistral-7B-Instruct-v0.1
+        cost_per_hour: 0.06
+
+  - name: Cohere
+    class: llmproxy.provider.cohere.cohere.Cohere
+    models:
+      - name: command
+        cost_per_token_input: 0.0000015
+        cost_per_token_output: 0.000002
+      - name: command-light
+        cost_per_token_input: 0.0000015
+        cost_per_token_output: 0.000002
+      - name: command-nightly
+        cost_per_token_input: 0.0000015
+        cost_per_token_output: 0.000002
+      - name: command-light-nightly
+        cost_per_token_input: 0.0000015
+        cost_per_token_output: 0.000002
+
+  - name: Vertexai
+    class: llmproxy.provider.google.vertexai.VertexAI
+    models:
+      - name: text-bison
+        cost_per_token_input: 0.0000015
+        cost_per_token_output: 0.000002
+      # Currently not working
+      # - name: chat-bison-32k
+      #   cost_per_token_input: 0.000002
+      #   cost_per_token_output: 0.000002
