+++ llmproxy/provider/base.py
@@ -2,7 +2,6 @@ from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import Any
 
-import tiktoken
 from llmproxy.utils.log import logger
