+++ llmproxy/provider/openai/chatgpt.py
@@ -1,4 +1,4 @@
-from llmproxy.models.base import BaseModel, CompletionResponse
+from llmproxy.provider.base import BaseModel, CompletionResponse
 from llmproxy.utils.enums import BaseEnum
 from llmproxy.utils.log import logger
 import openai
