+++ test.py
@@ -6,7 +6,10 @@ from llmproxy.llmproxy import LLMProxy
 def main() -> None:
     # prompt = "I am a man, not a man, but not a man, that is an apple, or a banana!"
     prompt = "What is the square root of i?"
-    proxy_client = LLMProxy(path_to_configuration="api_configuration.yml")
+    proxy_client = LLMProxy(
+        path_to_user_configuration="api_configuration.yml",
+        path_to_dev_configuration="developer_configuration.yml",
+    )
 
     output = proxy_client.route(route_type="cost", prompt=prompt)
     print(output)
